Project-Onboarding/requirements.txt
pyyaml == 6.0.1
requests == 2.31.0
argparse == 3 0.0.0
openpyxl == 3.1.2
